Modularization: move Bitcoin tests to the bitcoin family
