+++ libs/ledger-live-common/src/families/bitcoin/wallet-btc/__tests__/wallet.estimateMaxSpendable.test.ts
@@ -1,7 +1,7 @@
-import { DerivationModes } from "../../../../families/bitcoin/wallet-btc/types";
-import BitcoinLikeWallet from "../../../../families/bitcoin/wallet-btc/wallet";
-import * as utils from "../../../../families/bitcoin/wallet-btc/utils";
-import { Account } from "../../../../families/bitcoin/wallet-btc/account";
+import { DerivationModes } from "../types";
+import BitcoinLikeWallet from "../wallet";
+import * as utils from "../utils";
+import { Account } from "../account";
 
 describe("testing estimateMaxSpendable", () => {
   const wallet = new BitcoinLikeWallet();
