+++ libs/ledger-live-common/src/families/bitcoin/wallet-btc/__tests__/xpub.syncing.integration.test.ts
@@ -3,11 +3,11 @@
 // @ts-ignore
 import coininfo from "coininfo";
 import { toMatchFile } from "jest-file-snapshot";
-import { DerivationModes } from "../../../../families/bitcoin/wallet-btc/types";
-import BitcoinLikeStorage from "../../../../families/bitcoin/wallet-btc/storage";
-import BitcoinLikeExplorer from "../../../../families/bitcoin/wallet-btc/explorer";
-import Xpub from "../../../../families/bitcoin/wallet-btc/xpub";
-import * as currency from "../../../../families/bitcoin/wallet-btc/crypto";
+import { DerivationModes } from "../types";
+import BitcoinLikeStorage from "../storage";
+import BitcoinLikeExplorer from "../explorer";
+import Xpub from "../xpub";
+import * as currency from "../crypto";
 
 expect.extend({ toMatchFile });
 describe("xpub integration sync", () => {
