+++ libs/ledger-live-common/src/families/bitcoin/wallet-btc/__tests__/xpub.txs.litecoin.integration.test.ts
@@ -4,28 +4,29 @@ import * as bitcoin from "bitcoinjs-lib";
 import coininfo from "coininfo";
 import axios from "axios";
 import BigNumber from "bignumber.js";
-import { DerivationModes } from "../../../../families/bitcoin/wallet-btc/types";
-import Xpub from "../../../../families/bitcoin/wallet-btc/xpub";
-import Crypto from "../../../../families/bitcoin/wallet-btc/crypto/bitcoin";
-import BitcoinLikeExplorer from "../../../../families/bitcoin/wallet-btc/explorer";
-import BitcoinLikeStorage from "../../../../families/bitcoin/wallet-btc/storage";
-import { Merge } from "../../../../families/bitcoin/wallet-btc/pickingstrategies/Merge";
-import * as utils from "../../../../families/bitcoin/wallet-btc/utils";
+import { DerivationModes } from "../types";
+import Xpub from "../xpub";
+import Litecoin from "../crypto/litecoin";
+import BitcoinLikeExplorer from "../explorer";
+import BitcoinLikeStorage from "../storage";
+import { Merge } from "../pickingstrategies/Merge";
 
 const sleep = (ms: number) => new Promise((resolve) => setTimeout(resolve, ms));
 
+/*
+In order to launch this litecoin test locally
+TICKER=ltc TAG=latest LOG_LEVEL=debug docker-compose -f ./environments/explorer-praline-btc.yml up -d
+*/
 // FIXME Skipped because Praline required on CI
-describe.skip("testing xpub segwit transactions", () => {
-  const network = coininfo.bitcoin.regtest.toBitcoinJS();
-
+describe.skip("testing xpub legacy litecoin transactions", () => {
   const explorer = new BitcoinLikeExplorer({
     explorerURI: "http://localhost:20000/blockchain/v3",
     explorerVersion: "v3",
     disableBatchSize: true, // https://ledgerhq.atlassian.net/browse/BACK-2191
   });
-  const crypto = new Crypto({
-    network,
-  });
+
+  const network = coininfo.litecoin.test.toBitcoinJS();
+  const crypto = new Litecoin({ network });
 
   const xpubs = [1, 2, 3].map((i) => {
     const storage = new BitcoinLikeStorage();
@@ -41,7 +42,7 @@ describe.skip("testing xpub segwit transactions", () => {
       explorer,
       crypto,
       xpub: node.neutered().toBase58(),
-      derivationMode: DerivationModes.SEGWIT,
+      derivationMode: DerivationModes.LEGACY,
     });
 
     return {
@@ -81,15 +82,14 @@ describe.skip("testing xpub segwit transactions", () => {
   });
 
   let expectedFee1: number;
-  it("should send a 1 btc tx to xpubs[1].xpub", async () => {
+  it("should send a 1 litecoin tx to xpubs[1].xpub", async () => {
     const { address } = await xpubs[1].xpub.getNewAddress(0, 0);
     const changeAddress = await xpubs[0].xpub.getNewAddress(1, 0);
-
     const psbt = new bitcoin.Psbt({ network });
 
     const utxoPickingStrategy = new Merge(
       xpubs[0].xpub.crypto,
-      DerivationModes.SEGWIT,
+      xpubs[0].xpub.derivationMode,
       []
     );
 
@@ -103,37 +103,14 @@ describe.skip("testing xpub segwit transactions", () => {
         sequence: 0,
       });
 
-    inputs.forEach((input, i) => {
+    inputs.forEach((input) => {
+      const nonWitnessUtxo = Buffer.from(input.txHex, "hex");
       const tx = bitcoin.Transaction.fromHex(input.txHex);
-      const keyPair = xpubs[0].signer(
-        associatedDerivations[i][0],
-        associatedDerivations[i][1]
-      );
-      const p2wpkh = bitcoin.payments.p2wpkh({
-        pubkey: keyPair.publicKey,
-        network,
-      });
-      const p2sh = bitcoin.payments.p2sh({ redeem: p2wpkh, network });
-      const redeemScript = p2sh.redeem?.output?.toString("hex");
-      const publickeyHash = bitcoin.crypto
-        .ripemd160(bitcoin.crypto.sha256(keyPair.publicKey))
-        .toString("hex");
 
       psbt.addInput({
         hash: tx.getId(),
         index: input.output_index,
-        witnessUtxo: {
-          script: Buffer.from(
-            `a914${bitcoin.crypto
-              .hash160(Buffer.from(`0014${publickeyHash}`, "hex"))
-              .toString("hex")}87`,
-            "hex"
-          ),
-          value: Number(input.value),
-        },
-        redeemScript: redeemScript
-          ? Buffer.from(redeemScript, "hex")
-          : undefined,
+        nonWitnessUtxo,
       });
     });
     outputs.forEach((output) => {
@@ -155,6 +132,7 @@ describe.skip("testing xpub segwit transactions", () => {
     });
     psbt.finalizeAllInputs();
     const rawTxHex = psbt.extractTransaction().toHex();
+    //
 
     try {
       await xpubs[0].xpub.broadcastTx(rawTxHex);
@@ -162,7 +140,10 @@ describe.skip("testing xpub segwit transactions", () => {
       // eslint-disable-next-line no-console
       console.log("broadcast error", e);
     }
-    await sleep(10000);
+
+    // time for explorer to sync
+    await sleep(40000);
+
     try {
       const { address: mineAddress } = await xpubs[2].xpub.getNewAddress(0, 0);
       await axios.post(`http://localhost:28443/chain/mine/${mineAddress}/1`);
@@ -170,26 +151,18 @@ describe.skip("testing xpub segwit transactions", () => {
       // eslint-disable-next-line no-console
       console.log("praline error");
     }
-
-    // time for explorer to sync
     await sleep(40000);
-
     await xpubs[0].xpub.sync();
     await xpubs[1].xpub.sync();
 
     expectedFee1 =
-      utils.maxTxSizeCeil(
-        inputs.length,
-        outputs.map((o) => o.address),
-        false,
-        crypto,
-        DerivationModes.SEGWIT
-      ) * 100;
+      10 * 100 + inputs.length * 100 * 180 + outputs.length * 34 * 100;
+
     expect((await xpubs[0].xpub.getXpubBalance()).toNumber()).toEqual(
       5700000000 - 100000000 - expectedFee1
     );
     expect((await xpubs[1].xpub.getXpubBalance()).toNumber()).toEqual(
       100000000
     );
-  }, 180000);
+  }, 100000);
 });
